fixed minor issuews on map

+++ loagma_crm/lib/screens/admin/enhanced_attendance_management_screen.dart
@@ -663,14 +663,14 @@ class _EnhancedAttendanceManagementScreenState
                       Expanded(
                         child: _buildStatItemDouble(
                           'Avg Hours',
-                          avgWorkHours,
+                          avgWorkHours.toDouble(),
                           Colors.indigo,
                         ),
                       ),
                       Expanded(
                         child: _buildStatItemDouble(
                           'Total Hours',
-                          totalWorkHours,
+                          totalWorkHours.toDouble(),
                           Colors.teal,
                         ),
                       ),
